Move source files to new src folder
