+++ src/items/metadata/package.json
