+++ src/rules/condition-builder.js
@@ -1,5 +1,5 @@
 const operations = require('./operation-builder');
-const items = require('../items');
+const items = require('../items/items');
 
 /**
  * Condition that wraps a function to determine whether if passes
