+++ src/rules/package.json
