+++ src/rules/trigger-builder.js
